Create Linear search in Python
This code is a simple linear search in Python which takes in an input, i.e., length of an input array, and the elements of the array. Once the array is created, it displays the array, and takes in an enter to be found. It displays the position of the element in the array (list), else it displays "element not found"